Fix HTML escaping
Better avoidance of node render recursion

Updated approvals

Add memoisation

Update approvals
