+++ templates/html/node/StringNode.phtml
@@ -1,0 +1,1 @@
+<?php echo htmlspecialchars($this->nodePrinter->print($object, [])) ?>
